requirements.txt
Flask==0.11.1
Jinja2==2.8
MarkupSafe==0.23
PyYAML==3.12
Werkzeug==0.11.11
click==6.6
itsdangerous==0.24
pymongo==3.3.1
pytz==2016.7
